@@ -9,19 +9,25 @@
     }
   },
   "require": {
-    "php": ">=5.6",
+    "php": ">=7.1",
     "guzzlehttp/guzzle": "^6.2.0",
     "doctrine/cache": "^1.6.0",
     "vlucas/phpdotenv": "^2.2.0",
-    "zendframework/zend-hydrator": "^2.2.1"
+    "zendframework/zend-hydrator": "^2.2.1",
+    "ext-json": "*"
   },
   "require-dev": {
     "satooshi/php-coveralls": "^1",
-    "phpunit/phpunit": "^5.3",
-    "squizlabs/php_codesniffer": "^2",
+    "phpunit/phpunit": "^7.3",
+    "squizlabs/php_codesniffer": "^3.3",
     "phing/phing": "^2",
-    "facile-it/paraunit": "dev-master",
-    "symfony/var-dumper": "^3.1.6"
+    "facile-it/paraunit": "^0.12",
+    "phpstan/phpstan": "^0.10",
+    "phpstan/phpstan-deprecation-rules": "^0.10",
+    "phpstan/phpstan-phpunit": "^0.10",
+    "phpstan/phpstan-symfony": "^0.10",
+    "symfony/var-dumper": "^4.1",
+    "jangregor/phpstan-prophecy": "^0.2.0"
   },
   "scripts": {
     "post-install-cmd": [
@@ -35,5 +41,6 @@
     "bin-dir": "bin"
   },
   "extra": {},
-  "prefer-stable": true
+  "prefer-stable": true,
+  "description": "An OOP PHP implementation of appnexus API client"
 }
