@@ -7,21 +7,18 @@ use Audiens\AppnexusClient\CacheableInterface;
 use Audiens\AppnexusClient\entity\SegmentBilling;
 use Audiens\AppnexusClient\exceptions\RepositoryException;
 use Doctrine\Common\Cache\Cache;
-use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
 
-/**
- * Class SegmentBillingRepository
- */
 class SegmentBillingRepository implements CacheableInterface
 {
     use CachableTrait;
 
-    const BASE_URL = 'https://api.adnxs.com/segment-billing-category';
+    public const BASE_URL         = 'https://api.adnxs.com/segment-billing-category';
+    public const SANDBOX_BASE_URL = 'http://api-test.adnxs.com/segment-billing-category';
+    public const CACHE_NAMESPACE  = 'appnexus_segment_billing_repository_find_all';
+    public const CACHE_EXPIRATION = 3600;
 
-    const SANDBOX_BASE_URL = 'http://api-test.adnxs.com/segment-billing-category';
-
-    /** @var Client */
+    /** @var ClientInterface */
     protected $client;
 
     /** @var  int */
@@ -33,63 +30,38 @@ class SegmentBillingRepository implements CacheableInterface
     /** @var  string */
     protected $baseUrl;
 
-    const CACHE_NAMESPACE = 'appnexus_segment_billing_repository_find_all';
-
-    const CACHE_EXPIRATION = 3600;
-
-    /**
-     * SegmentRepository constructor.
-     *
-     * @param ClientInterface $client
-     * @param Cache|null      $cache
-     */
-    public function __construct(ClientInterface $client, Cache $cache = null)
+    public function __construct(ClientInterface $client, Cache $cache, int $memberId)
     {
-        $this->client = $client;
-        $this->cache = $cache;
-        $this->cacheEnabled = $cache instanceof Cache;
-        $this->baseUrl = self::BASE_URL;
+        $this->client   = $client;
+        $this->cache    = $cache;
+        $this->memberId = $memberId;
+        $this->baseUrl  = self::BASE_URL;
     }
 
-    /**
-     * @return string
-     */
-    public function getBaseUrl()
+    public function getBaseUrl(): string
     {
         return $this->baseUrl;
     }
 
-    /**
-     * @param string $baseUrl
-     */
     public function setBaseUrl($baseUrl)
     {
         $this->baseUrl = $baseUrl;
     }
 
-
-    /**
-     * @param SegmentBilling $segmentBilling
-     *
-     * @return RepositoryResponse
-     * @throws RepositoryException
-     */
-    public function add(SegmentBilling $segmentBilling)
+    public function add(SegmentBilling $segmentBilling): RepositoryResponse
     {
-
-        $compiledUrl = $this->baseUrl.'?member_id='.$segmentBilling->getMemberId();
+        $compiledUrl = $this->baseUrl.'?member_id='.$this->memberId;
 
         $payload = [
             'segment-billing-category' => $segmentBilling->toArray(),
         ];
-
 
         $response = $this->client->request('POST', $compiledUrl, ['body' => json_encode($payload)]);
 
         $repositoryResponse = RepositoryResponse::fromResponse($response);
 
         if ($repositoryResponse->isSuccessful()) {
-            $stream = $response->getBody();
+            $stream          = $response->getBody();
             $responseContent = json_decode($stream->getContents(), true);
 
             $stream->rewind();
@@ -108,20 +80,13 @@ class SegmentBillingRepository implements CacheableInterface
         return $repositoryResponse;
     }
 
-    /**
-     * @param SegmentBilling $segmentBilling
-     *
-     * @return RepositoryResponse
-     * @throws RepositoryException
-     */
-    public function update(SegmentBilling $segmentBilling)
+    public function update(SegmentBilling $segmentBilling): RepositoryResponse
     {
-
         if (!$segmentBilling->getId()) {
             throw RepositoryException::missingSegmentBillingId($segmentBilling);
         }
 
-        $compiledUrl = $this->baseUrl.'?member_id='.$segmentBilling->getMemberId();
+        $compiledUrl = $this->baseUrl.'?member_id='.$this->memberId;
 
         $payload = [
             'segment-billing-category' => $segmentBilling->toArray(),
@@ -129,22 +94,12 @@ class SegmentBillingRepository implements CacheableInterface
 
         $response = $this->client->request('PUT', $compiledUrl, ['body' => json_encode($payload)]);
 
-        $repositoryResponse = RepositoryResponse::fromResponse($response);
-
-        return $repositoryResponse;
+        return RepositoryResponse::fromResponse($response);
     }
 
-    /**
-     * @param $memberId
-     * @param $segmentId
-     * @return SegmentBilling
-     * @throws RepositoryException
-     */
-    public function findOneBySegmentId($memberId, $segmentId)
+    public function findOneBySegmentId($segmentId): ?SegmentBilling
     {
-
-        $compiledUrl = $this->baseUrl.'?member_id='.$memberId.'&segment_id='.$segmentId;
-
+        $compiledUrl = $this->baseUrl.'?member_id='.$this->memberId.'&segment_id='.$segmentId;
 
         $response = $this->client->request('GET', $compiledUrl);
 
@@ -154,7 +109,7 @@ class SegmentBillingRepository implements CacheableInterface
             throw RepositoryException::failed($repositoryResponse);
         }
 
-        $stream = $response->getBody();
+        $stream          = $response->getBody();
         $responseContent = json_decode($stream->getContents(), true);
         $stream->rewind();
 
@@ -166,32 +121,18 @@ class SegmentBillingRepository implements CacheableInterface
             throw RepositoryException::genericFailed('Expected only one results. Found '.count($responseContent['response']['segment-billing-categories']));
         }
 
-
         return SegmentBilling::fromArray($responseContent['response']['segment-billing-categories'][0]);
     }
 
-    /**
-     * @param     $memberId
-     * @param int $start
-     * @param int $maxResults
-     *
-     * @return SegmentBilling[]|null
-     * @throws RepositoryException
-     */
-    public function findAll($memberId, $start = 0, $maxResults = 100)
+    public function findAll($start = 0, $maxResults = 100): ?array
     {
+        $cacheKey = self::CACHE_NAMESPACE.sha1($this->memberId.$start.$maxResults);
 
-        $cacheKey = self::CACHE_NAMESPACE.sha1($memberId.$start.$maxResults);
-
-
-        if ($this->isCacheEnabled()) {
-            if ($this->cache->contains($cacheKey)) {
-                return $this->cache->fetch($cacheKey);
-            }
+        if ($this->cache->contains($cacheKey)) {
+            return $this->cache->fetch($cacheKey);
         }
 
-        $compiledUrl = $this->baseUrl.'?member_id='.$memberId.'&start_element='.$start.'&num_elements='.$maxResults;
-
+        $compiledUrl = $this->baseUrl.'?member_id='.$this->memberId.'&start_element='.$start.'&num_elements='.$maxResults;
 
         $response = $this->client->request('GET', $compiledUrl);
 
@@ -201,10 +142,9 @@ class SegmentBillingRepository implements CacheableInterface
             throw RepositoryException::failed($repositoryResponse);
         }
 
-        $stream = $response->getBody();
+        $stream          = $response->getBody();
         $responseContent = json_decode($stream->getContents(), true);
         $stream->rewind();
-
 
         $result = [];
 
@@ -216,28 +156,17 @@ class SegmentBillingRepository implements CacheableInterface
             $result[] = SegmentBilling::fromArray($segmentBillingArray);
         }
 
-        if ($this->isCacheEnabled()) {
-            $this->cache->save($cacheKey, $result, self::CACHE_EXPIRATION);
-        }
+        $this->cache->save($cacheKey, $result, self::CACHE_EXPIRATION);
 
         return $result;
     }
 
-    /**
-     * @param $memberId
-     * @param $id
-     *
-     * @return RepositoryResponse
-     */
-    public function remove($memberId, $id)
+    public function remove($id)
     {
-
-        $compiledUrl = $this->baseUrl.'?member_id='.$memberId.'&id='.$id;
+        $compiledUrl = $this->baseUrl.'?member_id='.$this->memberId.'&id='.$id;
 
         $response = $this->client->request('DELETE', $compiledUrl);
 
-        $repositoryResponse = RepositoryResponse::fromResponse($response);
-
-        return $repositoryResponse;
+        return RepositoryResponse::fromResponse($response);
     }
 }
