@@ -7,22 +7,19 @@ use Audiens\AppnexusClient\CacheableInterface;
 use Audiens\AppnexusClient\entity\Segment;
 use Audiens\AppnexusClient\exceptions\RepositoryException;
 use Doctrine\Common\Cache\Cache;
-use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
 
-/**
- * Class SegmentRepository
- */
 class SegmentRepository implements CacheableInterface
 {
 
     use CachableTrait;
 
-    const BASE_URL = 'https://api.adnxs.com/segment/';
+    public const BASE_URL         = 'https://api.adnxs.com/segment/';
+    public const SANDBOX_BASE_URL = 'http://api-test.adnxs.com/segment/';
+    public const CACHE_NAMESPACE  = 'appnexus_segment_repository_find_all';
+    public const CACHE_EXPIRATION = 3600;
 
-    const SANDBOX_BASE_URL = 'http://api-test.adnxs.com/segment/';
-
-    /** @var Client */
+    /** @var ClientInterface */
     protected $client;
 
     /** @var  int */
@@ -34,51 +31,27 @@ class SegmentRepository implements CacheableInterface
     /** @var  string */
     protected $baseUrl;
 
-    const CACHE_NAMESPACE = 'appnexus_segment_repository_find_all';
-
-    const CACHE_EXPIRATION = 3600;
-
-    /**
-     * SegmentRepository constructor.
-     *
-     * @param ClientInterface $client
-     * @param Cache|null      $cache
-     */
-    public function __construct(ClientInterface $client, Cache $cache = null)
+    public function __construct(ClientInterface $client, Cache $cache, int $memberId)
     {
-        $this->client = $client;
-        $this->cache = $cache;
-        $this->cacheEnabled = $cache instanceof Cache;
-        $this->baseUrl = self::BASE_URL;
+        $this->client   = $client;
+        $this->cache    = $cache;
+        $this->memberId = $memberId;
+        $this->baseUrl  = self::BASE_URL;
     }
 
-    /**
-     * @return string
-     */
-    public function getBaseUrl()
+    public function getBaseUrl(): string
     {
         return $this->baseUrl;
     }
 
-    /**
-     * @param string $baseUrl
-     */
     public function setBaseUrl($baseUrl)
     {
         $this->baseUrl = $baseUrl;
     }
 
-
-    /**
-     * @param Segment $segment
-     *
-     * @return RepositoryResponse
-     * @throws RepositoryException
-     */
-    public function add(Segment $segment)
+    public function add(Segment $segment): RepositoryResponse
     {
-
-        $compiledUrl = $this->baseUrl.$segment->getMemberId();
+        $compiledUrl = $this->baseUrl.$this->memberId;
 
         $payload = [
             'segment' => $segment->toArray(),
@@ -89,11 +62,11 @@ class SegmentRepository implements CacheableInterface
         $repositoryResponse = RepositoryResponse::fromResponse($response);
 
         if ($repositoryResponse->isSuccessful()) {
-            $stream = $response->getBody();
+            $stream          = $response->getBody();
             $responseContent = json_decode($stream->getContents(), true);
             $stream->rewind();
 
-            if (!(isset($responseContent['response']['segment']['id']))) {
+            if (!isset($responseContent['response']['segment']['id'])) {
                 throw RepositoryException::wrongFormat(serialize($responseContent));
             }
 
@@ -103,38 +76,22 @@ class SegmentRepository implements CacheableInterface
         return $repositoryResponse;
     }
 
-    /**
-     * @param $memberId
-     * @param $id
-     *
-     * @return RepositoryResponse
-     */
-    public function remove($memberId, $id)
+    public function remove($id): RepositoryResponse
     {
-
-        $compiledUrl = $this->baseUrl.$memberId.'/'.$id;
+        $compiledUrl = $this->baseUrl.$this->memberId.'/'.$id;
 
         $response = $this->client->request('DELETE', $compiledUrl);
 
-        $repositoryResponse = RepositoryResponse::fromResponse($response);
-
-        return $repositoryResponse;
+        return RepositoryResponse::fromResponse($response);
     }
 
-    /**
-     * @param Segment $segment
-     *
-     * @return RepositoryResponse
-     * @throws RepositoryException
-     */
-    public function update(Segment $segment)
+    public function update(Segment $segment): RepositoryResponse
     {
-
         if (!$segment->getId()) {
             throw RepositoryException::missingId($segment);
         }
 
-        $compiledUrl = $this->baseUrl.$segment->getMemberId().'/'.$segment->getId();
+        $compiledUrl = $this->baseUrl.$this->memberId.'/'.$segment->getId();
 
         $payload = [
             'segment' => $segment->toArray(),
@@ -142,21 +99,12 @@ class SegmentRepository implements CacheableInterface
 
         $response = $this->client->request('PUT', $compiledUrl, ['body' => json_encode($payload)]);
 
-        $repositoryResponse = RepositoryResponse::fromResponse($response);
-
-        return $repositoryResponse;
+        return RepositoryResponse::fromResponse($response);
     }
 
-    /**
-     * @param $id
-     * @param $memberId
-     *
-     * @return Segment|null
-     */
-    public function findOneById($memberId, $id)
+    public function findOneById($id): ?Segment
     {
-
-        $compiledUrl = $this->baseUrl.$memberId.'/'.$id;
+        $compiledUrl = $this->baseUrl.$this->memberId.'/'.$id;
 
         $response = $this->client->request('GET', $compiledUrl);
 
@@ -166,33 +114,22 @@ class SegmentRepository implements CacheableInterface
             return null;
         }
 
-        $stream = $response->getBody();
+        $stream          = $response->getBody();
         $responseContent = json_decode($stream->getContents(), true);
         $stream->rewind();
 
         return Segment::fromArray($responseContent['response']['segment']);
     }
 
-    /**
-     * @param     $memberId
-     * @param int $start
-     * @param int $maxResults
-     *
-     * @return Segment[]|null
-     * @throws RepositoryException
-     */
-    public function findAll($memberId, $start = 0, $maxResults = 100)
+    public function findAll($start = 0, $maxResults = 100): ?array
     {
+        $cacheKey = self::CACHE_NAMESPACE.sha1($this->memberId.$start.$maxResults);
 
-        $cacheKey = self::CACHE_NAMESPACE.sha1($memberId.$start.$maxResults);
-
-        if ($this->isCacheEnabled()) {
-            if ($this->cache->contains($cacheKey)) {
-                return $this->cache->fetch($cacheKey);
-            }
+        if ($this->cache->contains($cacheKey)) {
+            return $this->cache->fetch($cacheKey);
         }
 
-        $compiledUrl = $this->baseUrl.$memberId."?start_element=$start&num_elements=$maxResults";
+        $compiledUrl = $this->baseUrl.$this->memberId."?start_element=$start&num_elements=$maxResults";
 
         $response = $this->client->request('GET', $compiledUrl);
 
@@ -202,7 +139,7 @@ class SegmentRepository implements CacheableInterface
             throw RepositoryException::failed($repositoryResponse);
         }
 
-        $stream = $response->getBody();
+        $stream          = $response->getBody();
         $responseContent = json_decode($stream->getContents(), true);
         $stream->rewind();
 
@@ -216,9 +153,7 @@ class SegmentRepository implements CacheableInterface
             $result[] = Segment::fromArray($segmentArray);
         }
 
-        if ($this->isCacheEnabled()) {
-            $this->cache->save($cacheKey, $result, self::CACHE_EXPIRATION);
-        }
+        $this->cache->save($cacheKey, $result, self::CACHE_EXPIRATION);
 
         return $result;
     }
